requirements.txt
# requirements.txt
streamlit
google-generativeai
streamlit-extras